fix: rutas varias
