+++ client/Front_PrograWeb/src/pages/AdminPages/ListOrders.jsx
@@ -32,3 +32,5 @@ export const ListaOrdenes = () =>{
         </div>
     );
 }
+
+export default ListaOrdenes;
